Extract StorageConfig outside ServiceConfiguration
