+++ src/main/java/uk/gov/hmcts/reform/blobrouter/config/ServiceConfiguration.java
@@ -6,7 +6,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import javax.validation.constraints.NotNull;
 
 @ConfigurationProperties(prefix = "service")
 public class ServiceConfiguration {
@@ -23,56 +22,4 @@ public class ServiceConfiguration {
             .collect(Collectors.toMap(StorageConfig::getName, Function.identity()));
     }
 
-    public static class StorageConfig {
-        private String name;
-        private int sasValidity;
-
-        @NotNull
-        private TargetStorageAccount targetStorageAccount;
-
-        @NotNull
-        private String targetContainer;
-
-        private boolean isEnabled = true;
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public int getSasValidity() {
-            return sasValidity;
-        }
-
-        public void setSasValidity(int sasValidity) {
-            this.sasValidity = sasValidity;
-        }
-
-        public boolean isEnabled() {
-            return isEnabled;
-        }
-
-        public void setEnabled(boolean enabled) {
-            isEnabled = enabled;
-        }
-
-        public TargetStorageAccount getTargetStorageAccount() {
-            return targetStorageAccount;
-        }
-
-        public void setTargetStorageAccount(TargetStorageAccount targetStorageAccount) {
-            this.targetStorageAccount = targetStorageAccount;
-        }
-
-        public String getTargetContainer() {
-            return targetContainer;
-        }
-
-        public void setTargetContainer(String targetContainer) {
-            this.targetContainer = targetContainer;
-        }
-    }
 }
