+++ src/main/java/uk/gov/hmcts/reform/blobrouter/tasks/processors/BlobProcessor.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.blobrouter.config.ServiceConfiguration;
+import uk.gov.hmcts.reform.blobrouter.config.StorageConfig;
 import uk.gov.hmcts.reform.blobrouter.config.TargetStorageAccount;
 import uk.gov.hmcts.reform.blobrouter.data.EnvelopeRepository;
 import uk.gov.hmcts.reform.blobrouter.data.model.NewEnvelope;
@@ -48,7 +49,7 @@ public class BlobProcessor {
     private final BlobSignatureVerifier signatureVerifier;
 
     // container-specific configuration, by container name
-    private final Map<String, ServiceConfiguration.StorageConfig> storageConfig;
+    private final Map<String, StorageConfig> storageConfig;
 
     public BlobProcessor(
         @Qualifier("storage-client") BlobServiceClient storageClient,
@@ -105,7 +106,7 @@ public class BlobProcessor {
                 boolean validSignature = signatureVerifier.verifyZipSignature(blobName, rawBlob);
 
                 if (validSignature) {
-                    ServiceConfiguration.StorageConfig containerConfig = storageConfig.get(containerName);
+                    StorageConfig containerConfig = storageConfig.get(containerName);
                     TargetStorageAccount targetStorageAccount = containerConfig.getTargetStorageAccount();
                     var targetContainerName = containerConfig.getTargetContainer();
