+++ src/test/java/uk/gov/hmcts/reform/blobrouter/tasks/processors/BlobProcessorTest.java
@@ -11,6 +11,7 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.blobrouter.config.ServiceConfiguration;
+import uk.gov.hmcts.reform.blobrouter.config.StorageConfig;
 import uk.gov.hmcts.reform.blobrouter.config.TargetStorageAccount;
 import uk.gov.hmcts.reform.blobrouter.data.EnvelopeRepository;
 import uk.gov.hmcts.reform.blobrouter.data.model.NewEnvelope;
@@ -312,7 +313,7 @@ class BlobProcessorTest {
         String targetContainer,
         TargetStorageAccount targetStorageAccount
     ) {
-        var containerConfig = new ServiceConfiguration.StorageConfig();
+        var containerConfig = new StorageConfig();
         containerConfig.setEnabled(true);
         containerConfig.setName(sourceContainer);
         containerConfig.setTargetContainer(targetContainer);
